Add files via upload
The alt_final.py file is the program for the nucleotide sequence trimmer for a FASTQ file. It includes the prompt given to create the program and comments on what the reasonings for each function are. The assignment2v2.fastq is the file used to write and test my code.